limited x axis rotation, can spawn on any rotation
I limited the rotation of the camera on the x axis so the player can't view under the grid. It might be a little limited atm, but can change. you can now spawn the blocks on any rotation relatively well. There is a bug I noticed: if you put a tile onto the grid and throw it in the trash, then spawn the same tile and don't
place it on the grid, it will snap to the location of the one that was thrown away.
