+++ Temp3D_BYN_Project/Assets/Scripts/CameraMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e428294754677454a87fb5a29df0b802
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
