+++ Temp3D_BYN_Project/Library/StateCache/SceneView/b2/b26cafd87359696c420111f92939153d.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":false,"pivot":{"x":471.99737548828127,"y":622.2095947265625,"z":21.8931884765625},"rotation":{"x":-0.0566600002348423,"y":-0.005680728238075972,"z":-0.00031785780447535217,"w":0.9983859062194824},"size":309.7845764160156,"orthographic":false}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":false,"pivot":{"x":-8.39942741394043,"y":-0.04807038977742195,"z":-0.14691640436649323},"rotation":{"x":-0.12704616785049439,"y":0.7152150273323059,"z":-0.13480398058891297,"w":-0.6739403605461121},"size":5.675459861755371,"orthographic":false}
